break resizing feedback loop
the feedback loop is this: the client send the server a Resize request;
the server resizes the pane, then send the client a TabResized event;
the client starts a resync, which calls ClientPane::resize;
ClientPane::resize sends the server another Resize request.

add a parameter to Pane::resize to stop Resize request from being sent
in cases like this.